mk802: Add support for the original mk802 (512MB RAM version)
Add support for the original mk802 board with 512MB RAM, this is the mk802
with the female mini-hdmi connector, where as the mk802ii has a male normal
hdmi connecter sticking out of it.

There have been some reports on the sunxi-list that the original mk802 does
not work, because the kernel tries to load the axp20 pmu driver and this
fails because the mk802 does not have a pmu.

This is likely caused by using the script.bin with which newer batches of the
original mk802 ship. Which seems to be a copy of an mk802ii script.bin, and
the [pmu] section contains "pmu_used=1" causing the driver to wrongly get
loaded. The mk802.fex file in the sunxi-boards repository is correct and
with that my mk802 without pmu (which comes with the wrong script.bin in its
nand) works fine.

Signed-off-by: Hans de Goede <hdegoede@redhat.com>