sun5i: Default console to uart0 (v2)
With support for A10s board coming we're going to have a lot of boards
with the console on uart0, so we're going to have a lot of
CONS_INDEX=x statements in boards.cfg anyways.

Given this it makes sense to just default to uart0, since that is a more
logical default and also unifies the console default with sun4i.

Chanes in v2: Don't set CONS_INDEX=2 for sun5i_sdcon.

Signed-off-by: Hans de Goede <hdegoede@redhat.com>
Signed-off-by: Henrik Nordstrom <henrik@henriknordstrom.net>