sunxi: Enable pull-up on uart rx pin
Many boards have the uart pins hooked up to test points, or jumper pins for
debugging use, so most of the time they end up floating.

Sometimes this results in reception of spurious messages, and when this
happens when u-boot is waiting for a keypress to go interactive, the boot
will hang without any indication why (the u-boot messages are not seen as
nothing is hooked up to the uart, otherwise we would not have the problem).

To fix this this commit enables the build-in pull-up on the rx pin of the
uart. This actually brings the u-boot kernel closer to how the kernel does
things. The kernel uses the fex file contents to determine wether or not
to enable the pull up. And in all fex files for recent boards (those based
on evb-1.2 or newer) the pull-up us enabled for both the rx and tx pins.

I see little value in enabling the pull-up for the tx pins, but enabling it
on the rx pins fixes one of my allwinner devices from booting almost never,
to booting just fine, all of the time.

Signed-off-by: Hans de Goede <hdegoede@redhat.com>