Less magic for the watchdog timer
This started out as a cleanup of the wdt to remove some magic values but
turned into somewhat of a rewrite and relocate patch.

The only big change and 'gotcha' is that you actually use a watchdog
time of 0 now. Allwinner made 0 to be 0.5 seconds, the fastest followed
by 1 for etc. upto 16 seconds, the maximum. There are some gaps in the
allwinner tables, those timings are rounded up to nearest second up. To
disable the WDT anything above 16 suffices.

Signed-off-by: Oliver Schinagl <oliver@schinagl.nl>