sun5i: Tweak pll6 init value
There are multiple ways to get 600 MHz from PLL6, the sun5i ccmu is very
similar to the sun4i and sun7i ccmu.

This commit changes the PLL6 initialization we do for sun5i to make the PLL6
reg value match that of sun4i and sun7i. This also makes it closer to the
sun5i power on default, as we are now only changing the K-factor from 3 to 1.

Signed-off-by: Hans de Goede <hdegoede@redhat.com>