sunxi: Use SPDX license headers.
Following eca3aeb352c9 "Licenses: introduce SPDX Unique Lincense Identifiers"
switch sunxi files to use SPDX license identifiers. As with that changeset,
this patch is supposed to be only a change to the textual representation of the
license information, but in no way any change to the actual license terms. With
this patch applied, all files will still be licensed under the same terms they
were before.

The vast majority of these were done by re-applying a single diff hunk, the
rest were done by resolving the resulting reject file, using emacs diff mode,
while checking that the mismatch wasn't due to a license difference.

The set of files was determined with:
  git grep 'This program is free software' | grep sunxi
  git diff v2014.01  | lsdiff --strip 1 | xargs grep 'This program is free software'

This may or may not be a prerequisite to upstreaming, but I figured I may as
well clean it up.

Signed-off-by: Ian Campbell <ijc@hellion.org.uk>