sunxi: use clrsetbit in gpio/pinmux code
I was concerned that this would remove some barriers which were previously
implicitly present in the readl/writel calls.

Comparing the preprocessed code shows that the old code had compiler barrers
(not CPU instruction barriers) between the reads and writes while the new code
does everything as a single statement without barriers. I think this is safe.

In any case the generated code is identical (confirmed with objdump).

Signed-off-by: Ian Campbell <ijc@hellion.org.uk>
Signed-off-by: Hans de Goede <hdegoede@redhat.com>