sunxi: avoid sr32 for APB1 clock setup.
In some cases we are setting all of the defined bits of the configuration
register, so there is no need for a read/modify/write via either sr32 or
clrsetbits, just write the value we need.

Tested on Cubietruck (sun7i) and compile tested only for Colombus (sun6i).

Having made these changes sr32 is now unused in sunxi so remove the prototype.

Signed-off-by: Ian Campbell <ijc@hellion.org.uk>
---
v2: - Correct sun6i clock source defines
    - Clearer version of clock_twi_onoff for both