sunxi: Add i12-tvbox board
The i12 tvbox is an A20 based android tvbox, with 1G RAM, 4G nand flash,
ap6210 sdio wifi + bt (broadcom BCM43362 sdio wifi + uart attached brcm bt),
2USB host ports using USB-A receptacles and a micro-usb receptacle for USB OTG,
and 100Mbit ethernet.

The PCB is labelled i12-a20 hence I've named the board i12-a20. It is used
in noname allwinner A20 tv-boxes, which are sometimes sold with Q5 or
QT840A as product name.

Note this patch does NOT enable the ethernet port, as the phy-power is
controller through a gpio-pin (PH21) and we don't have support for having
a gpio power pin for the phy (yet) in the GMAC driver (nor in the EMAC one).

Signed-off-by: Hans de Goede <hdegoede@redhat.com>